Backend/requirements.txt
scikit-learn==1.1.3
numpy==1.23.5
flask
flask-cors
python-docx
PyPDF2
